Fix evm family unit tests
