+++ libs/ledger-live-common/src/families/evm/__tests__/transaction.unit.test.ts
@@ -1,11 +1,12 @@
 import { ethers } from "ethers";
 import BigNumber from "bignumber.js";
-import { findCurrencyByTicker } from "../../../currencies";
-import * as API from "../api/rpc";
+import { findCryptoCurrencyById } from "../../../currencies";
+import { transactionToEthersTransaction } from "../adapters";
+import { makeAccount } from "../testUtils";
+import * as API from "../api/rpc.common";
 import {
   fromTransactionRaw,
   toTransactionRaw,
-  transactionToEthersTransaction,
   transactionToUnsignedTransaction,
 } from "../transaction";
 import {
@@ -16,9 +17,8 @@ import {
 } from "../types";
 
 const testData = Buffer.from("testBufferString").toString("hex");
-const fakeAccount = {
-  currency: { ...findCurrencyByTicker("ETH"), rpc: "fakeNodeRpc" },
-};
+const currency = findCryptoCurrencyById("ethereum")!;
+const fakeAccount = makeAccount("0xBob", currency);
 const rawEip1559Tx: EvmTransactionEIP1559Raw = {
   amount: "100",
   useAllAmount: false,
@@ -107,6 +107,7 @@ describe("EVM Family", () => {
     describe("transactionToEthersTransaction", () => {
       it("should build convert an EIP1559 ledger live transaction to an ethers transaction", () => {
         const ethers1559Tx: ethers.Transaction = {
+          from: "0xBob",
           to: "0xkvn",
           nonce: 0,
           gasLimit: ethers.BigNumber.from(21000),
@@ -118,11 +119,14 @@ describe("EVM Family", () => {
           maxPriorityFeePerGas: ethers.BigNumber.from(10000),
         };
 
-        expect(transactionToEthersTransaction(eip1559Tx)).toEqual(ethers1559Tx);
+        expect(transactionToEthersTransaction(eip1559Tx, fakeAccount)).toEqual(
+          ethers1559Tx
+        );
       });
 
       it("should build convert an legacy ledger live transaction to an ethers transaction", () => {
         const legacyEthersTx: ethers.Transaction = {
+          from: "0xBob",
           to: "0xkvn",
           nonce: 0,
           gasLimit: ethers.BigNumber.from(21000),
@@ -133,7 +137,7 @@ describe("EVM Family", () => {
           gasPrice: ethers.BigNumber.from(10000),
         };
 
-        expect(transactionToEthersTransaction(legacyTx)).toEqual(
+        expect(transactionToEthersTransaction(legacyTx, fakeAccount)).toEqual(
           legacyEthersTx
         );
       });
